# [improvement](fs_bench) optimize the usage of fs benchmark tool for hdfs