--- a/build.sh
+++ b/build.sh
@@ -637,6 +637,10 @@ EOF
         cp -r -p "${DORIS_HOME}/be/output/lib/debug_info" "${DORIS_OUTPUT}/be/lib"/
     fi
 
+    if [[ "${BUILD_FS_BENCHMARK}" = "ON" ]]; then
+        cp -r -p "${DORIS_HOME}/bin/run-fs-benchmark.sh" "${DORIS_OUTPUT}/be/bin/"/
+    fi
+
     extensions_modules=("")
     extensions_modules+=("java-udf")
     extensions_modules+=("jdbc-scanner")