@morningman 2023-06-25
run buildall