@morningman 2023-06-26
run buildall